refactored out market state and moved order processing in to processors chosen by a factory over type enum
