+++ TradeEngine/src/main/java/com/gala/sam/tradeEngine/domain/Order.java
@@ -1,7 +1,12 @@
 package com.gala.sam.tradeEngine.domain;
 
+import lombok.Getter;
+
+@Getter
 public class Order {
 
+    public enum OrderType {STOP, READY_LIMIT, READY_MARKET}
 
+    OrderType type;
 
 }
