+++ TradeEngine/src/main/java/com/gala/sam/tradeEngine/domain/dataStructures/MarketState.java
@@ -1,0 +1,36 @@
+package com.gala.sam.tradeEngine.domain.dataStructures;
+
+import com.gala.sam.tradeEngine.domain.ReadyOrder;
+import com.gala.sam.tradeEngine.domain.StopOrder;
+import com.gala.sam.tradeEngine.domain.Trade;
+
+import java.util.*;
+
+public class MarketState {
+
+    private List<Trade> trades = new ArrayList<>();
+    private Map<String, TickerData> tickerQueues = new TreeMap<>();
+    private List<StopOrder> stopOrders = new LinkedList<>();
+
+
+    public Map<String, TickerData> getTickerQueues() {
+        return tickerQueues;
+    }
+
+    public List<StopOrder> getStopOrders() {
+        return stopOrders;
+    }
+
+    public List<Trade> getTrades() {
+        return trades;
+    }
+
+    public TickerData getTickerQueueGroup(ReadyOrder marketOrder) {
+        TickerData queues = tickerQueues.get(marketOrder.getTicker());
+        if (queues == null) {
+            queues = new TickerData();
+            tickerQueues.put(marketOrder.getTicker(), queues);
+        }
+        return queues;
+    }
+}
