+++ TradeEngine/src/main/java/com/gala/sam/tradeEngine/utils/OrderProcessor/OrderProcessor.java
@@ -1,0 +1,7 @@
+package com.gala.sam.tradeEngine.utils.OrderProcessor;
+
+import com.gala.sam.tradeEngine.domain.Order;
+
+public interface OrderProcessor {
+    <T extends Order> void process(T order);
+}
